Add Activators.
