+++ src/jp.ac.nagoya_u.is.nces.a_rte.codegen/META-INF/MANIFEST.MF
@@ -22,3 +22,4 @@ Import-Package: com.google.common.base;version="[11.0.0,15.0.0)",
  com.google.common.collect;version="[11.0.0,15.0.0)",
  com.google.common.io;version="[11.0.0,15.0.0)",
  com.google.common.util.concurrent;version="[11.0.0,15.0.0)"
+Bundle-Activator: jp.ac.nagoya_u.is.nces.a_rte.codegen.internal.Activator
