+++ src/jp.ac.nagoya_u.is.nces.a_rte.m2m/META-INF/MANIFEST.MF
@@ -7,8 +7,10 @@ Bundle-Vendor: %providerName
 Bundle-Localization: plugin
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-ActivationPolicy: lazy
-Require-Bundle: org.eclipse.emf.ecore;bundle-version="2.9.0",
+Require-Bundle: org.eclipse.core.runtime;bundle-version="3.10.0",
+ org.eclipse.emf.ecore;bundle-version="2.9.0",
  org.eclipse.emf.query;bundle-version="1.7.0",
  com.google.guava;bundle-version="11.0.2",
  jp.ac.nagoya_u.is.nces.a_rte.model;bundle-version="1.2.2"
 Export-Package: jp.ac.nagoya_u.is.nces.a_rte.m2m
+Bundle-Activator: jp.ac.nagoya_u.is.nces.a_rte.m2m.internal.Activator
