+++ src/jp.ac.nagoya_u.is.nces.a_rte.m2m/src/jp/ac/nagoya_u/is/nces/a_rte/m2m/internal/Activator.java
@@ -1,0 +1,35 @@
+package jp.ac.nagoya_u.is.nces.a_rte.m2m.internal;
+
+import org.eclipse.core.runtime.Plugin;
+import org.osgi.framework.BundleActivator;
+import org.osgi.framework.BundleContext;
+
+public class Activator extends Plugin implements BundleActivator {
+
+	// The plug-in ID
+	public static final String PLUGIN_ID = "jp.pizzafactory.a_rte.m2m"; //$NON-NLS-1$
+
+	// The shared instance
+	private static Activator plugin;
+
+	@Override
+	public void start(BundleContext context) throws Exception {
+		super.start(context);
+		plugin = this;
+	}
+
+	@Override
+	public void stop(BundleContext context) throws Exception {
+		plugin = null;
+		super.stop(context);
+	}
+
+	/**
+	 * Returns the shared instance
+	 *
+	 * @return the shared instance
+	 */
+	public static Activator getDefault() {
+		return plugin;
+	}
+}
