+++ src/jp.ac.nagoya_u.is.nces.a_rte.model/META-INF/MANIFEST.MF
@@ -28,9 +28,9 @@ Export-Package: jp.ac.nagoya_u.is.nces.a_rte.model,
  jp.ac.nagoya_u.is.nces.a_rte.model.rte_test.impl,
  jp.ac.nagoya_u.is.nces.a_rte.model.rte_test.util,
  jp.ac.nagoya_u.is.nces.a_rte.model.util
-Require-Bundle: org.eclipse.emf.query.ocl;bundle-version="2.0.0",
+Require-Bundle: org.eclipse.core.runtime;bundle-version="3.10.0",
+ org.eclipse.emf.query.ocl;bundle-version="2.0.0",
  com.google.guava;bundle-version="11.0.2",
- org.eclipse.core.runtime,
  org.eclipse.emf.ecore,
  org.eclipse.osgi
 Import-Package: com.google.common.base;version="[11.0.0,15.0.0)",
@@ -39,3 +39,4 @@ Import-Package: com.google.common.base;version="[11.0.0,15.0.0)",
  com.google.common.io;version="[14.0.0,15.0.0)",
  com.google.common.util.concurrent;version="[11.0.0,15.0.0)"
 Bundle-ActivationPolicy: lazy
+Bundle-Activator: jp.ac.nagoya_u.is.nces.a_rte.model.internal.Activator
