+++ src/jp.ac.nagoya_u.is.nces.a_rte.validation/META-INF/MANIFEST.MF
@@ -10,10 +10,11 @@ Bundle-Localization: plugin
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Export-Package: jp.ac.nagoya_u.is.nces.a_rte.validation,
  jp.ac.nagoya_u.is.nces.a_rte.validation.internal
-Require-Bundle: org.eclipse.ocl.ecore;bundle-version="3.3.0",
+Require-Bundle: org.eclipse.core.runtime;bundle-version="3.10.0",
+ org.eclipse.ocl.ecore;bundle-version="3.3.0",
  org.eclipse.emf.query.ocl;bundle-version="2.0.0",
  com.google.guava;bundle-version="11.0.2",
- org.eclipse.emf.ecore,
- jp.ac.nagoya_u.is.nces.a_rte.model;bundle-version="1.2.2",
- org.eclipse.osgi
+ org.eclipse.emf.ecore;bundle-version="2.10.2",
+ jp.ac.nagoya_u.is.nces.a_rte.model;bundle-version="1.2.2"
 Bundle-ActivationPolicy: lazy
+Bundle-Activator: jp.ac.nagoya_u.is.nces.a_rte.validation.internal.Activator
